sui_indexer_alt_consistent_store::restore

Function start_restorer

pub async fn start_restorer(
    path: impl AsRef<Path>,
    formal_snapshot_args: FormalSnapshotArgs,
    connection_args: StorageConnectionArgs,
    restore_args: RestoreArgs,
    pipelines: BTreeSet<String>,
    config: DbConfig,
    registry: &Registry,
) -> Result<(Service, Finalizer)>

Set-up and run the Restorer, using the provided arguments (expected to be extracted from the command-line). The service will run until restoration complete.

path is the path to the RocksDB database to restore into. It will be created if it does not exist. formal_snapshot_args describes the formal snapshot source, connection_args controls how to connect to it, restore_args controls the restoration process itself, and config configures the RocksDB database.