sui_indexer_alt_consistent_store

Function start_service

pub async fn start_service(
    path: impl AsRef<Path>,
    indexer_args: IndexerArgs,
    client_args: ClientArgs,
    rpc_args: RpcArgs,
    version: &'static str,
    config: ServiceConfig,
    registry: &Registry,
) -> Result<Service>

Set-up and run the Indexer and RPC service, using the provided arguments (expected to be extracted from the command-line).

path is the path to the RocksDB database,which will be created if it does not exist. indexer_args and client_args control the behavior of the Indexer, while rpc_args controls the behavior of the RPC service.

The service spins up auxiliary services (to expose metrics, run the indexer, and the RPC), and will clean these up on shutdown as well.